[add]新旧見比べられるよう準備
